requirements_test.txt
# Test requirements for Home Assistant integration
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-homeassistant-custom-component>=0.13.0
pypentairthermalwifi
